🔨 fix capitalisation and formatting
