+++ frontend/src/components/documentation/Documentation.jsx
@@ -1,9 +1,7 @@
 // UCL API navbar
-import 'Styles/navbar.scss'
-
 import { NavBar } from 'Layout/Items.jsx'
 import React from 'react'
-
+import 'Styles/navbar.scss'
 import GetInvolved from './GetInvolved/GetInvolved.jsx'
 import Welcome from './Meta/Welcome.jsx'
 import Authorise from './Routes/OAuth/Authorise.jsx'
@@ -24,95 +22,93 @@ import GetDataDepartments from './Routes/Timetable/GetDataDepartments.jsx'
 import GetDataModules from './Routes/Timetable/GetDataModules.jsx'
 import GetPersonalTimetable from './Routes/Timetable/GetPersonalTimetable.jsx'
 import GetTimetableByModules from './Routes/Timetable/GetTimetableByModules.jsx'
-import WorkspacesGetImage from './Routes/Workspaces/GetImage.jsx'
-import WorkspacesGetLastSensorUpdate from './Routes/Workspaces/GetLastSensorUpdate.jsx'
-import WorkspacesGetLiveImage from './Routes/Workspaces/GetLiveImage.jsx'
-import WorkspacesGetHistoricalTimeData from './Routes/Workspaces/GetSensorHistoricalTimeData.jsx'
-import WorkspacesGetSensors from './Routes/Workspaces/GetSensors.jsx'
-import WorkspaceGetSensorsSummary from './Routes/Workspaces/GetSensorSummary.jsx'
-import WorkspacesGetSurveys from './Routes/Workspaces/GetSurveys.jsx'
+import {
+  GetImage as WorkspacesGetImage,
+  GetLastSensorUpdate as WorkspacesGetLastSensorUpdate,
+  GetLiveImage as WorkspacesGetLiveImage,
+  GetSensorHistoricalTimeData as WorkspacesGetHistoricalTimeData,
+  GetSensors as WorkspacesGetSensors,
+  GetSensorSummary as WorkspaceGetSensorsSummary,
+  GetSurveys as WorkspacesGetSurveys,
+} from './Routes/Workspaces'
 import SectionHeader from './SectionHeader.jsx'
-import Sidebar from './Sidebar.jsx'
+import Sidebar from './Sidebar'
 
-export default class DocumentationComponent extends React.Component {
+const DocumentationComponent = () => (
+  <>
+    <NavBar isScroll={false} />
+    <Sidebar />
+    <div className="main">
+      <Welcome key={`Welcome`} />
 
-  render() {
-    return (
-      <div>
-        <NavBar isScroll={false} />
-        <Sidebar />
-        <div className="main">
-          <Welcome key={`Welcome`} />
+      <SectionHeader
+        key={`SectionHeaderWelcome`}
+        link="oauth"
+        title="OAuth"
+      />
+      <OAuthIntro key={`OAuthIntro`} />
+      <Authorise key={`Authorise`} />
+      <Token key={`Token`} />
+      <UserData key={`UserData`} />
+      <StudentNumber key={`StudentNumber`} />
 
-          <SectionHeader
-            key={`SectionHeaderWelcome`}
-            link="oauth"
-            title="OAuth"
-          />
-          <OAuthIntro key={`OAuthIntro`} />
-          <Authorise key={`Authorise`} />
-          <Token key={`Token`} />
-          <UserData key={`UserData`} />
-          <StudentNumber key={`StudentNumber`} />
+      <SectionHeader
+        link="roombookings"
+        title="Room Bookings"
+        key={`SectionHeaderRoomBookings`}
+      />
+      <GetRooms key={`GetRooms`} />
+      <GetBookings key={`GetBookings`} />
+      <GetEquiment key={`GetEquiment`} />
+      <GetFreeRooms key={`GetFreeRooms`} />
+      <Webhooks key={`Webhooks`} />
 
-          <SectionHeader
-            link="roombookings"
-            title="Room Bookings"
-            key={`SectionHeaderRoomBookings`}
-          />
-          <GetRooms key={`GetRooms`} />
-          <GetBookings key={`GetBookings`} />
-          <GetEquiment key={`GetEquiment`} />
-          <GetFreeRooms key={`GetFreeRooms`} />
-          <Webhooks key={`Webhooks`} />
+      <SectionHeader
+        link="search"
+        title="Search"
+        key={`SectionHeaderSearch`}
+      />
+      <GetPeople key={`GetPeople`} />
 
-          <SectionHeader
-            link="search"
-            title="Search"
-            key={`SectionHeaderSearch`}
-          />
-          <GetPeople key={`GetPeople`} />
+      <SectionHeader
+        link="timetable"
+        title="Timetable"
+        key={`SectionHeaderTimetable`}
+      />
+      <GetPersonalTimetable key={`GetPersonalTimetable`} />
+      <GetTimetableByModules key={`GetTimetableByModules`} />
+      <GetDataDepartments key={`GetDataDepartments`} />
+      <GetDataModules key={`GetDataModules`} />
+      <GetDataCourses key={`GetDataCourses`} />
+      <GetDataCoursesModules key={`GetDataCoursesModules`} />
 
-          <SectionHeader
-            link="timetable"
-            title="Timetable"
-            key={`SectionHeaderTimetable`}
-          />
-          <GetPersonalTimetable key={`GetPersonalTimetable`} />
-          <GetTimetableByModules key={`GetTimetableByModules`} />
-          <GetDataDepartments key={`GetDataDepartments`} />
-          <GetDataModules key={`GetDataModules`} />
-          <GetDataCourses key={`GetDataCourses`} />
-          <GetDataCoursesModules key={`GetDataCoursesModules`} />
+      <SectionHeader
+        link="resources"
+        title="Resources"
+        key={`SectionHeaderResources`}
+      />
+      <DesktopAvailability key={`DesktopAvailability`} />
 
-          <SectionHeader
-            link="resources"
-            title="Resources"
-            key={`SectionHeaderResources`}
-          />
-          <DesktopAvailability key={`DesktopAvailability`} />
+      <SectionHeader
+        link="workspaces"
+        title="Workspaces"
+        key={`SectionHeaderWorkspaces`}
+      />
+      <WorkspacesGetSurveys key={`WorkspacesGetSurveys`} />
+      <WorkspacesGetSensors key={`WorkspacesGetSensors`} />
+      <WorkspacesGetHistoricalTimeData
+        key={`WorkspacesGetHistoricalTimeData`}
+      />
+      <WorkspacesGetLastSensorUpdate
+        key={`WorkspacesGetLastSensorUpdate`}
+      />
+      <WorkspaceGetSensorsSummary key={`WorkspaceGetSensorsSummary`} />
+      <WorkspacesGetImage key={`WorkspacesGetImage`} />
+      <WorkspacesGetLiveImage key={`WorkspacesGetLiveImage`} />
 
-          <SectionHeader
-            link="workspaces"
-            title="Workspaces"
-            key={`SectionHeaderWorkspaces`}
-          />
-          <WorkspacesGetSurveys key={`WorkspacesGetSurveys`} />
-          <WorkspacesGetSensors key={`WorkspacesGetSensors`} />
-          <WorkspacesGetHistoricalTimeData
-            key={`WorkspacesGetHistoricalTimeData`}
-          />
-          <WorkspacesGetLastSensorUpdate
-            key={`WorkspacesGetLastSensorUpdate`}
-          />
-          <WorkspaceGetSensorsSummary key={`WorkspaceGetSensorsSummary`} />
-          <WorkspacesGetImage key={`WorkspacesGetImage`} />
-          <WorkspacesGetLiveImage key={`WorkspacesGetLiveImage`} />
+      <GetInvolved key={`GetInvolved`} />
+    </div>
+  </>
+)
 
-          <GetInvolved key={`GetInvolved`} />
-        </div>
-      </div>
-    )
-  }
-
-}
+export default DocumentationComponent
