+++ frontend/src/components/documentation/Routes/Workspaces/index.js
@@ -1,0 +1,10 @@
+export { default as GetImage } from './GetImage.jsx'
+export { default as GetLastSensorUpdate } from './GetLastSensorUpdate.jsx'
+export { default as GetLiveImage } from './GetLiveImage.jsx'
+export {
+  default as GetSensorHistoricalTimeData,
+} from './GetSensorHistoricalTimeData.jsx'
+export { default as GetSensors } from './GetSensors.jsx'
+export { default as GetSensorSummary } from './GetSensorSummary.jsx'
+export { default as GetSurveys } from './GetSurveys.jsx'
+
