+++ frontend/src/components/documentation/index.js
@@ -1,0 +1,2 @@
+export { default } from './Documentation.jsx'
+
