+++ frontend/src/sass/documentation.scss
@@ -158,7 +158,6 @@ code {
 
 .section-header {
   border-bottom: 1px solid #e4e4e4;
-  border-top: 1px solid #e4e4e4;
   font-size: 36px;
   font-weight: 600;
   margin-bottom: 0;
